% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{ml1_LN_rep}
\alias{ml1_LN_rep}
\title{MEGH I}
\usage{
ml1_LN_rep(par, times, status, des, des_t, n.clust)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{n.clust}{: number of clusters}

\item{par:}{all model parameters (reparameterised)}
}
\value{
Marginal log likelihood with Log-Normal baseline hazard (reparameterised)
}
\description{
MEGH I
}
