% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{ml1Ind_gamma_rep}
\alias{ml1Ind_gamma_rep}
\title{MEGH I}
\usage{
ml1Ind_gamma_rep(par, times, status, des, des_t, index)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{index}{: cluster ID}

\item{par:}{all model parameters (reparameterised)}
}
\value{
Individual Marginal log likelihood with Gamma baseline hazard (reparameterised)
}
\description{
MEGH I
}
