% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{ml2_gamma}
\alias{ml2_gamma}
\title{MEGH II}
\usage{
ml2_gamma(par, times, status, des, des_t, n.clust)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{n.clust}{: number of clusters}

\item{par:}{all model parameters in the original parameterisation}
}
\value{
Marginal log likelihood with Gamma baseline hazard (original parameterisation)
}
\description{
MEGH II
}
