% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{MSurvPGW1t}
\alias{MSurvPGW1t}
\title{MEGH I. Student-t random effects}
\usage{
MSurvPGW1t(t, par, des, des_t, index, NMC)
}
\arguments{
\item{des}{: Design matrix}

\item{des_t}{: Design matrix for time-dependent effects}

\item{index}{: individual ID}

\item{NMC}{: number of Monte Carlo samples to be used to produce the marginal}

\item{t:}{time}

\item{par:}{Marginal maximum likelihood estimator (original parameterisation)}
}
\value{
The marginal survival function for cluster "index" at time t with PGW baseline hazard
}
\description{
MEGH I. Student-t random effects
}
