% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{MSurvPGW1tpn}
\alias{MSurvPGW1tpn}
\title{MEGH I. twopiece normal random effects}
\usage{
MSurvPGW1tpn(t, par, des, des_t, index, NMC)
}
\arguments{
\item{des}{: Design matrix}

\item{des_t}{: Design matrix for time-dependent effects}

\item{index}{: individual ID}

\item{NMC}{: number of Monte Carlo samples to be used to produce the marginal}

\item{t:}{time}

\item{par:}{Marginal maximum likelihood estimator (original parameterisation)}
}
\value{
Marginal survival function with PGW baseline hazard for cluster "index" at time t
}
\description{
MEGH I. twopiece normal random effects
}
