% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{MSurvPGW2}
\alias{MSurvPGW2}
\title{MEGH II.
Marginal survival function for cluster "index" at time t}
\usage{
MSurvPGW2(t, par, des, des_t, index, NMC)
}
\arguments{
\item{des}{: Design matrix}

\item{des_t}{: Design matrix for time-dependent effects}

\item{index}{: individual ID}

\item{NMC}{: number of Monte Carlo samples to be used to produce the marginal}

\item{t:}{time}

\item{par:}{Marginal maximum likelihood estimator (original parameterisation)}
}
\value{
The marginal survival function for cluster "index" at time t with PGW baseline hazard
}
\description{
MEGH II.
Marginal survival function for cluster "index" at time t
}
