% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{simMEGH}
\alias{simMEGH}
\title{Function to simulate times to event from a model with MEGH structure.}
\usage{
simMEGH(
  seed,
  des = NULL,
  des_t = NULL,
  ID,
  alpha,
  beta,
  theta,
  Sigma,
  cens,
  restr,
  distr
)
}
\arguments{
\item{seed}{: seed for simulation}

\item{des}{: Design matrix for hazard level effects}

\item{des_t}{: Design matrix for time-dependent effects}

\item{ID}{: Individual identifiers}

\item{beta}{: regression parameters multiplying the hazard}

\item{theta}{: parameters of the baseline hazard}

\item{Sigma}{: Standard deviation of the random effects or covariance matrix}

\item{cens}{: Censoring times}

\item{restr}{: random effects structure ("I", "II", "GENERAL")}

\item{distr}{: baseline hazard distribution ("LN", "LL", "PGW" or "gamma")}

\item{n}{: sample size}
}
\value{
a list containing the survival times, vital status, and ID
}
\description{
Function to simulate times to event from a model with MEGH structure.
}
