% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{simMEGH_t}
\alias{simMEGH_t}
\title{Function to simulate times to event from a model with MEGH structure and Student-t random effects}
\usage{
simMEGH_t(
  seed,
  des = NULL,
  des_t = NULL,
  ID,
  alpha,
  beta,
  theta,
  sigma,
  nu,
  cens,
  restr,
  distr
)
}
\arguments{
\item{seed}{: seed for simulation}

\item{des}{: Design matrix}

\item{des_t}{: Design matrix for time-dependent effects}

\item{ID}{: Individual identifiers}

\item{beta}{: regression parameters multiplying the hazard}

\item{sigma}{: Standard deviation of the random effects}

\item{nu}{: degrees of freedom}

\item{cens}{: Censoring times}

\item{restr}{: random effects structure ("I", "II" or "III")}

\item{distr}{: baseline hazard distribution ("LN", "LL", "PGW" or "gamma")}

\item{theta:}{parameters of the baseline hazard}

\item{n}{: sample size}
}
\value{
a list containing the survival times, vital status, and ID
}
\description{
Function to simulate times to event from a model with MEGH structure and Student-t random effects
}
