% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{simMEGH_TPN}
\alias{simMEGH_TPN}
\title{Function to simulate times to event from a model with MEGH structure and twopiece normal random effects}
\usage{
simMEGH_TPN(
  seed,
  des = NULL,
  des_t = NULL,
  ID,
  alpha,
  beta,
  theta,
  sigma,
  skew,
  cens,
  restr,
  distr
)
}
\arguments{
\item{seed}{: seed for simulation}

\item{des}{: Design matrix for hazard-level effects}

\item{des_t}{: Design matrix for time-dependent effects}

\item{ID}{: Individual identifiers}

\item{beta}{: regression parameters multiplying the hazard}

\item{theta}{: parameters of the baseline hazard}

\item{sigma}{: Standard deviation of the random effects}

\item{skew}{: skewness parameter (eps parameterisation) of the random effects}

\item{cens}{: Censoring times}

\item{restr}{: random effects structure ("I", "II" or "III")}

\item{distr}{: baseline hazard distribution ("LN", "LL", "PGW" or "gamma")}

\item{n}{: sample size}
}
\value{
list of survival times, vital status, and ID.
}
\description{
Function to simulate times to event from a model with MEGH structure and twopiece normal random effects
}
