% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{tml1_gamma_rep}
\alias{tml1_gamma_rep}
\title{MEGH I. Student-t random effects.}
\usage{
tml1_gamma_rep(par, times, status, des, des_t, n.clust)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{par:}{all model parameters reparameterised to the real line}
}
\value{
Marginal log likelihood with Gamma baseline hazard (reparameterised)
}
\description{
MEGH I. Student-t random effects.
}
