% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{tml2_LL_rep}
\alias{tml2_LL_rep}
\title{MEGH II. Student-t random effects.}
\usage{
tml2_LL_rep(par, times, status, des, des_t, n.clust)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{par:}{all model parameters reparameterised to the real line}
}
\value{
Marginal log likelihood with Log-logistic baseline hazard (reparameterised)
}
\description{
MEGH II. Student-t random effects.
}
