% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{tml2Ind_LL_rep}
\alias{tml2Ind_LL_rep}
\title{MEGH II. Student-t random effects.}
\usage{
tml2Ind_LL_rep(par, times, status, des, des_t, index)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{index}{: cluster ID}

\item{par:}{all model parameters (reparameterised)}
}
\value{
Individual Marginal log likelihood with Log-Logistic baseline hazard (reparameterised)
}
\description{
MEGH II. Student-t random effects.
}
