% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{tpnml1_PGW_rep}
\alias{tpnml1_PGW_rep}
\title{MEGH I. twopiece normal random effects.}
\usage{
tpnml1_PGW_rep(par, times, status, des, des_t, n.clust)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix for hazard-level effects (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{n.clust}{: number of clusters}

\item{par:}{all model parameters (reparameterised)}
}
\value{
Marginal log likelihood with PGW baseline hazard (reparameterised)
}
\description{
MEGH I. twopiece normal random effects.
}
