% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{tpnml1Ind_LN_rep}
\alias{tpnml1Ind_LN_rep}
\title{MEGH I. twopiece normal random effects.}
\usage{
tpnml1Ind_LN_rep(par, times, status, des, des_t, index)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{index}{: cluster ID}

\item{par:}{all model parameters (reparameterised)}
}
\value{
Individual Marginal log-likelihood with Log-Normal baseline hazard (reparameterised)
}
\description{
MEGH I. twopiece normal random effects.
}
