% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{tpnml2_LL_rep}
\alias{tpnml2_LL_rep}
\title{MEGH II. twopiece normal random effects.}
\usage{
tpnml2_LL_rep(par, times, status, des, des_t, n.clust)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix for hazard-level effects (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{n.clust}{: number of clusters}

\item{par:}{all model parameters (reparameterised)}
}
\value{
Marginal log likelihood with Log-logistic baseline hazard (reparameterised)
}
\description{
MEGH II. twopiece normal random effects.
}
