% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{tpnml2Ind_LN}
\alias{tpnml2Ind_LN}
\title{MEGH II. twopiece normal random effects.}
\usage{
tpnml2Ind_LN(par, times, status, des, des_t, index)
}
\arguments{
\item{times}{: survival times}

\item{status}{: vital status}

\item{des}{: design matrix (p x n)}

\item{des_t}{: design matrix for time-dependent effects (q x n)}

\item{index}{: cluster ID}

\item{par:}{all model parameters in the original parameterisation}
}
\value{
Individual Marginal log-likelihood with Log-Normal baseline hazard (original parameterisation)
}
\description{
MEGH II. twopiece normal random effects.
}
